connecting to mongodb
we did this because initially the server was starting first and then the mongodb was getting connected later on which might result to sum error. In the startServer function we are using await and async so that unless the db is getting connected, the server will not listen.
As initially there was no await or aysnc due to will the server was getting started earlier then mongodb because mongodb needs some time to connect and in that time the server used to start